Initial commit
Script that can handle A* Pathfinding for the gameobject the script is attached to, given a target gameobject. Currently no action after discovering path, as that will need to factor into how we do movement. But pathfinding is done.
 Note that it is functional, but currently lacks ability to mitigate what paths to check and consequently is inefficient. Unknown is this will be problematic or needs further review.
(Eg: If checks spot to the right, the next time around it will also check spot to left again so a "potential path" will be back and forth between two spots)