Branching as I modified a couple of common files: DungeonManager & Actor.
Interpolation is working between tiles, but there's some weird grid issues with drawing. No animation has been added yet, but it should be trivial after I get the timing down.
IDEA: Add a queued path to movement for the player. Every time they hit a direction, it gets enqueued, and thus the player can "build" a path so that the animation/interpolation of the player character may in fact "lag" behind and be concurrent with moving the character.